+++ examples/resources/meraki_appliance_l7_firewall_rules/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_l7_firewall_rules.example "<network_id>"