+++ docs/guides/changelog.md
@@ -13,6 +13,7 @@ description: |-
 - Add `meraki_appliance_inbound_cellular_firewall_rules` resource and data source
 - Add `meraki_appliance_inbound_firewall_rules` resource and data source
 - Add `meraki_appliance_l3_firewall_rules` resource and data source
+- Add `meraki_appliance_l7_firewall_rules` resource and data source
 
 ## 0.1.1